+++ packages/preview/web/src/ContentStateAnalyzer/ContentStateMetadata.spec.js
@@ -99,18 +99,21 @@ describe('content state media metadata', () => {
       height: 2800,
       type: 'image',
       url: '8bb438_e78b371c75ce42de8719dccfc97298a4.jpg',
+      id: '8bb438_e78b371c75ce42de8719dccfc97298a4.jpg',
       width: 4200,
     },
     {
       height: 1280,
       type: 'image',
       url: '8bb438_281af3d3281f4584a5a864c6c60f3a00.jpg',
+      id: '8bb438_281af3d3281f4584a5a864c6c60f3a00.jpg',
       width: 1920,
     },
     {
       height: 1081,
       type: 'image',
       url: '8bb438_0795e40ac4db438a8a723ea98dbeda10.jpg',
+      id: '8bb438_0795e40ac4db438a8a723ea98dbeda10.jpg',
       width: 1621,
     },
     {