+++ packages/editor-common/web/src/index.js
@@ -85,10 +85,11 @@ export {
   calculateDiff,
   getPostContentSummary,
   createSelection,
+  getBlockType,
+  setForceSelection,
 } from './Utils/draftUtils';
 export { isiOS } from './Utils/isiOS';
 export { getSelectionStyles } from './Utils/inlineStyleUtils';
-export { getConfigByFormFactor } from './Utils/getConfigByFormFactor';
 export { mergeToolbarSettings } from './Utils/mergeToolbarSettings';
 export {
   COMMANDS,