+++ packages/plugin-button/web/src/constants.js
@@ -1,9 +1,9 @@
+/* eslint-disable camelcase */
 export const BUTTON_TYPE = 'wix-draft-plugin-button';
 export const ALIGN_CENTER = 'center';
 export const settingsTabValue = 'settings';
 export const designTabValue = 'design';
 export const BUTTON_TYPES = ['primary', 'secondary'];
-export const DEFAULT_PALETTE = ['#FEFDFD', '#D5D4D4', '#000000', '#ABCAFF', '#81B0FF', '#0261FF'];
 export const DEFAULT_COLOR = '#000000';
 export const DEFAULT_SELECTION_COLOR = '#000000';
 export const COLOR_PICKER_TYPE = Object.freeze({
@@ -13,7 +13,7 @@ export const COLOR_PICKER_TYPE = Object.freeze({
 });
 
 export const COLORS = Object.freeze({
-  color1: '#FEFDFD',
+  color1: '#FFFFFF',
   color2: '#D5D4D4',
   color3: '#000000',
   color4: '#000000',
@@ -29,29 +29,6 @@ export const DEFAULT_CONFIG = {
   alignment: ALIGN_CENTER,
   size: 'small',
   width: 'fit-content',
-};
-
-export const getDefaultComponentData = (rel, target) => {
-  return {
-    config: DEFAULT_CONFIG,
-    button: {
-      settings: {
-        buttonText: 'Click Me',
-        url: '',
-        rel,
-        target,
-      },
-      design: {
-        activeButton: 0,
-        borderRadius: 0,
-        borderWidth: 0,
-        padding: 12,
-        background: COLORS.color8,
-        color: COLORS.color1,
-        borderColor: COLORS.color8,
-      },
-    },
-  };
 };
 
 export const buttonPreviews = colors => [