+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 ### :bug: Bug Fix
 - `preview`
   - [#903](https://github.com/wix-incubator/rich-content/pull/903) gallery fixed; resize flapping fixed
+- `gallery`
+  - [#903](https://github.com/wix-incubator/rich-content/pull/903) gallery size
 
 </details>
 <hr/>